+++ packages/react-components/react-card/src/components/Card/Card.test.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { render } from '@testing-library/react';
 import { Card } from './Card';
-import { isConformant } from '../../common/isConformant';
-import type { CardProps } from './Card.types';
+import { isConformant } from '../../testing/isConformant';
+import { CardProps } from './Card.types';
 
 describe('Card', () => {
   isConformant<CardProps>({